+++ repoze/who/plugins/friendlyform.py
@@ -59,7 +59,12 @@ class FriendlyFormPlugin(object):
 
     """
     implements(IChallenger, IIdentifier)
-
+    
+    classifications = {
+        IIdentifier: ["browser"],
+        IChallenger: ["browser"],
+        }
+    
     def __init__(self, login_form_url, login_handler_path, post_login_url,
                  logout_handler_path, post_logout_url, rememberer_name,
                  login_counter_name=None, charset="iso-8859-1"):